Add print/exit statements and improve shell link
The print and exit statements will be eventually implemented as
library functions for Zonkey. However, before I implement
functions and libraries in the interpreter I need to be able to test
the state of a script easily. Therefore, I have implemented these
functions as part of the Zonkey's grammar. These will then be replaced
when I start working on functions and libraries.

The shell and its link to the interpreter has also been improved so
the interpreter is kept alive between commands, whilst also allowing
the interpreter to communicate back to the shell (so exit() can work).
Allowing the interpreter to stay alive between commands is necessary to
allow the next steps of the interpreters development to be completed.