Add variable declarations and variable expressions
Allow the user to declare a variable, e.g. `Integer a = 1 - 2;` and then
use it later in the script, e.g. `print(a);`. At the moment, variables
can only have a global scope. However, globally declaring variables will
not be a feature in the eventual version of Zonkey, therefore, global
variables will be removed when I begin working on scope.